Use lazypush for queuing job of onLinksUpdateCompleted hook
This job that is being queued in every edit is being pushed directly.
This means in every edit, the appserver opens a request, communicate to
eventgate and then closes it, adding around 40ms to every edit.

You can see it in list of curl requests made in this xhgui report:
https://performance.wikimedia.org/xhgui/run/view?id=6153427113e6381f84c80315

It's not an urgent job and can be pushed later through deferred update
instead. The important thing is that it gets batched through all other
jobs being pushed.

Bug: T292048
Bug: T255502
Change-Id: Id9cad4e23c10751741e4b4779a42f38ff06cc668